pyproject.toml
[project]
name = "bootstrap"
version = "0.1.0"
description = "A robust, multi-platform system bootstrapper for Linux and macOS."
readme = "README.md"
authors = [
    { name = "Peter Prescott", email = "prescott.peter@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
bootstrap = "bootstrap:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
